requirements.txt
# Core Dependencies
boto3==1.34.144
pyyaml==6.0.1
python-dotenv==1.0.0
jsonlines==4.0.0

# PDF Processing
pymupdf==1.24.5
pypdf==4.2.0
Pillow==10.3.0

# Vector Database
chromadb==0.5.0
numpy<2.0  # ChromaDB compatibility

# Text Processing
tiktoken==0.7.0

# Utilities
tqdm==4.66.4
loguru==0.7.2

# Optional: For testing
pytest==8.2.2
pytest-asyncio==0.23.7
reportlab